requirements.txt
gradio
whisper
openai
python-dotenv
